Machine learning project
# Voice-Classification-using-ML
 We are building a  python application that analyses the sentiment behind the tone of the voice and predicts the Sentiment involved, we will use an open source dataset for the same .

The link to download the dataset used in the  implented model is given below.
https://www.kaggle.com/datasets/ejlok1/toronto-emotional-speech-set-tess

# Content of Dataset
There are a set of 200 target words were spoken in the carrier phrase "Say the word _' by 
two actresses (aged 26 and 64 years) and recordings were made of the set portraying each of 
seven emotions (anger, disgust, fear, happiness, pleasant surprise, sadness, and neutral). 
There are 2800 data points (audio files) in total.
The dataset is organised such that each of the two female actor and their emotions are 
contain within its own folder. And within that, all 200 target words audio file can be found. 
The format of the audio file is a WAV format

# Libraries used
• pandas 
• matplotlib
• keras
• tensorflow
• librosa
• numpy
• seaborn
• warnings

# Neural Network
LSTM Netwoork

# Accuracy of Model
The accuracy of the model is 68.21%